+++ src/subtitles/libssf/Subtitle.cpp
@@ -301,7 +301,7 @@ namespace ssf
 				CStringW direction = (*pDef)[L"direction"].IsValue() ? (*pDef)[L"direction"] : L"fw";
 				if(direction == L"fwbw" || direction == L"bwfw") t *= 2;
 
-				float n;
+				double n;
 				t = modf(t, &n);
 
 				if(direction == L"bw"
